+++ app/src/main/java/wien/mila/nachschlichten/data/local/entity/ArticleEntity.kt
@@ -9,7 +9,7 @@ data class ArticleEntity(
     val ean: String,
     val name: String,
     val unit: String,
-    val totalStock: Int,
+    val totalStock: Double,
     val price: Double,
     val lastSyncedAt: Long,
     val imagePath: String?
