+++ app/src/main/java/wien/mila/nachschlichten/domain/model/Article.kt
@@ -5,7 +5,7 @@ data class Article(
     val ean: String,
     val name: String,
     val unit: String,
-    val totalStock: Int,
+    val totalStock: Double,
     val price: Double,
     val lastSyncedAt: Long,
     val imagePath: String?
