+++ app/src/main/java/wien/mila/nachschlichten/domain/model/PendingItem.kt
@@ -6,7 +6,7 @@ data class PendingItem(
     val articleName: String,
     val articleEan: String,
     val shelfId: String,
-    val quantity: Int?,
+    val quantity: Double?,
     val createdAt: Long,
     val isDone: Boolean
 )
